Externalize composition modules
