+++ src/pykeen/nn/emb.py
@@ -10,16 +10,15 @@ import logging
 import warnings
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from typing import Any, Callable, ClassVar, Mapping, Optional, Sequence, Tuple, TypeVar, Union, cast
+from typing import Any, Callable, Mapping, Optional, Sequence, Tuple, TypeVar, Union, cast
 
 import numpy as np
 import torch
 import torch.nn
-from class_resolver import Resolver
 from torch import nn
 from torch.nn import functional
 
-from . import functional as pkf
+from .compositions import CompositionModule, composition_resolver
 from .init import init_phases, xavier_normal_, xavier_normal_norm_, xavier_uniform_, xavier_uniform_norm_
 from .message_passing import Decomposition, decomposition_resolver
 from .weighting import EdgeWeighting, SymmetricEdgeWeighting, edge_weight_resolver
@@ -583,59 +582,6 @@ class RGCNRepresentations(RepresentationModule):
         if indices is not None:
             x = x[indices]
         return x
-
-
-class CompositionModule(nn.Module, ABC):
-    """An (elementwise) composition function for vectors."""
-
-    @abstractmethod
-    def forward(self, a: torch.FloatTensor, b: torch.FloatTensor) -> torch.FloatTensor:
-        """Compose two batches of vectors.
-
-        The tensors have to be broadcastable.
-
-        :param a: shape: s_1
-            The first tensor.
-        :param b: shape: s_2
-            The second tensor.
-        :return: shape: s
-        """
-
-
-class FunctionalCompositionModule(CompositionModule):
-    """Composition by a function (i.e. state-less)."""
-
-    func: ClassVar[Callable[[torch.FloatTensor, torch.FloatTensor], torch.FloatTensor]]
-
-    def forward(self, a: torch.FloatTensor, b: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
-        return self.__class__.func(a, b)
-
-
-class SubtractionCompositionModule(FunctionalCompositionModule):
-    """Composition by element-wise subtraction."""
-
-    func = torch.sub
-
-
-class MultiplicationCompositionModule(FunctionalCompositionModule):
-    """Composition by element-wise multiplication."""
-
-    func = torch.mul
-
-
-class CircularCorrelationCompositionModule(FunctionalCompositionModule):
-    """Composition by circular correlation via :func:`pykeen.nn.functional.circular_correlation`."""
-
-    func = pkf.circular_correlation
-
-
-composition_resolver = Resolver.from_subclasses(
-    CompositionModule,
-    default=MultiplicationCompositionModule,
-    skip={
-        FunctionalCompositionModule,
-    },
-)
 
 
 class CompGCNLayer(nn.Module):
